Ben artifact perms (#76)
* Update package-lock

* Add auth perms to the following artifact routes: create, update, getAll, delete. Also removed artifact route to get a single artifact.

* Capitalized comments and removed a typo

* Updated a depreciated setting

* Update prettier config file

* Update permissions and auto format

* Update client requests to include token

* Add auth token to edit artifact request on the client

+++ server/controllers/artifact.js
@@ -1,4 +1,5 @@
 var Artifact = require('../models/artifact');
+var Circle = require('../models/circle');
 
 var createArtifact = async (req, res) => {
   try {
@@ -6,58 +7,83 @@ var createArtifact = async (req, res) => {
     delete req.body.id;
     // Create an artifact from the details in the request
     const artifact = new Artifact(req.body);
+    // Add the user's ID to the artifact objects uploader property
+    artifact.uploader = req.user.id;
     // Wait for the artifact to be saved in the database
     await artifact.save();
     // Return the artifact back to the client
     res.status(201).send(artifact.toObject());
   } catch (error) {
     // Return an error message as the artfact was not able to be created
-    res.status(400).send({error:'Unable to create artifact.'});
-  };
+    res.status(400).send({ error: 'Unable to create artifact.' });
+  }
 };
 
 var updateArtifact = async (req, res) => {
   try {
-    const query = {_id: req.params.id};
-    const updates = req.body;
-    await Artifact.updateOne(query, updates);
-    res.status(200).send();
+    const query = { _id: req.params.artifactId };
+    const updated = req.body;
+    // Get the requested artifact from the database
+    const artifact = await Artifact.findOne(query);
+    if ((artifact.uploader = req.user.id)) {
+      // If the user is the uploader, update the artifact as requested
+      await Artifact.updateOne(query, updated);
+      res.status(200).send();
+    } else {
+      // Otherwise don't and return an error response
+      res.status(400).send({ error: `You don't have permissions to update this artifact.` });
+    }
   } catch (error) {
-    res.status(400).send({error: 'Unable to update artifact.'});
-  };
+    res.status(400).send({ error: 'Unable to update artifact.' });
+  }
 };
 
-var getArtifact = async (req, res) => {
-  try {
-    const artifact = await Artifact.find({id:req.params.id});
-    res.status(200).send(artifacts);
-  } catch (error) {
-    res.status(400).send({error:'Unable to get that artifact.'})
-  };
+// Helper function to merge any number of arrays and remove duplicates
+var mergeArrays = (...arrays) => {
+  let jointArray = [];
+  arrays.forEach(array => {
+    jointArray = [...jointArray, ...array];
+  });
+  const uniqueArray = jointArray.filter((item, index) => jointArray.indexOf(item) === index);
+  return uniqueArray;
 };
 
 var getArtifacts = async (req, res) => {
   try {
-    const artifacts = await Artifact.find();
+    // Find all circles that the user is a member of
+    const circles = await Circle.find({ members: req.user.id });
+    // Find all artifact ids from within those circles (removing duplicates)
+    const artifactIds = mergeArrays(circles.map(circle => circle.artifacts));
+    // Get all artifacts with corresponding ids from the database
+    const artifacts = await Artifact.find({
+      $or: [{ _id: { $in: artifactIds } }, { uploader: req.user.id }],
+    });
     res.status(200).send(artifacts.map(artifact => artifact.toObject()));
   } catch (error) {
-    res.status(400).send({error:'Unable to get artifacts.'});
-  };
+    res.status(400).send({ error: 'Unable to get artifacts.' });
+  }
 };
 
 var deleteArtifact = async (req, res) => {
   try {
-    await Artifact.deleteOne({_id:req.params.id});
-    res.status(200).send();
+    // Get the requested artifact
+    const artifact = await Artifact.findOne({ _id: req.params.artifactId });
+    if (artifact.uploader === req.user.id) {
+      // If the user is the uploader then delete as requested
+      await Artifact.deleteOne({ _id: req.params.id });
+      res.status(200).send();
+    } else {
+      // Otherwise don't delete and return an error response
+      res.status(400).send({ error: `You don't have permissions to delete this artifact` });
+    }
   } catch (error) {
-    res.status(400).send({error:'Unable to delete artifact.'})
-  };
+    res.status(400).send({ error: 'Unable to delete artifact.' });
+  }
 };
 
 module.exports = {
-  createArtifact, 
+  createArtifact,
   updateArtifact,
   deleteArtifact,
-  getArtifact,
-  getArtifacts
+  getArtifacts,
 };
